+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added support for Woocommerce search widget @see [Duplicated search result](https://wordpress.org/support/topic/duplicated-search-result)
 * Fixed translation downloader tries to download woo translations for en_US locale
+* Fixed wrong product duplicate behavior 
 
 ###0.22